[AWIBOF-4306] Versioned segment context checkpoint implementation
Code changes:
added log message for assert cases. resolve code review comment.

+++ src/journal_manager/log_buffer/versioned_segment_ctx.cpp
@@ -221,7 +221,7 @@ VersionedSegmentCtx::_CheckLogGroupIdValidity(int logGroupId)
 void
 VersionedSegmentCtx::_CheckSegIdValidity(int segId)
 {
-    if (segId >= numSegments)
+    if (segId >= (int)numSegments)
     {
         POS_TRACE_ERROR((int)POS_EVENT_ID::JOURNAL_INVALID,
             "Failed to check segId validity, segId {} is invalid", segId);
